move rudra to keyoapp multisrc
