+++ src/en/rudrascans/build.gradle
@@ -1,7 +1,9 @@
 ext {
     extName = 'Rudra Scans'
     extClass = '.RudraScans'
-    extVersionCode = 1
+	themePkg = 'keyoapp'
+	baseUrl = 'https://rudrascans.com'
+    overrideVersionCode = 1
 }
 
 apply from: "$rootDir/common.gradle"
